# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyBox API",
    "description": "MyBox Application API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "file-controller"
        ],
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "files"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListFileUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/restore": {
      "post": {
        "tags": [
          "file-controller"
        ],
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/download": {
      "post": {
        "tags": [
          "file-controller"
        ],
        "operationId": "downloadFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/delete": {
      "post": {
        "tags": [
          "file-controller"
        ],
        "operationId": "delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/chunk/upload": {
      "post": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "uploadChunk",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UploadChunkRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUploadChunkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/chunk/merge": {
      "post": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "mergeChunks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChunkMergeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/chunk/init": {
      "post": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "initUpload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChunkInitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/chunk/resume/{uploadId}": {
      "get": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "resumeUpload",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChunkResumeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/chunk/progress/{uploadId}": {
      "get": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "getProgress",
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChunkProgressResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/chunk/cancel": {
      "delete": {
        "tags": [
          "chunk-upload-controller"
        ],
        "operationId": "cancelUpload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponseListFileUploadResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileUploadResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FileIdRequest": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "fileId"
        ]
      },
      "UploadChunkRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "chunkNumber": {
            "type": "integer",
            "format": "int32"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "chunkNumber",
          "file",
          "uploadId"
        ]
      },
      "ApiResponseUploadChunkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UploadChunkResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UploadChunkResponse": {
        "type": "object",
        "properties": {
          "chunkNumber": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "chunkHash": {
            "type": "string"
          }
        }
      },
      "ChunkMergeRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "uploadId"
        ]
      },
      "ApiResponseChunkMergeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ChunkMergeResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ChunkMergeResponse": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid"
          },
          "originalFileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "fileHash": {
            "type": "string"
          }
        }
      },
      "ChunkInitRequest": {
        "type": "object",
        "properties": {
          "originalFileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          }
        },
        "required": [
          "contentType",
          "fileSize",
          "originalFileName"
        ]
      },
      "ApiResponseChunkInitResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ChunkInitResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ChunkInitResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "chunkSize": {
            "type": "integer",
            "format": "int64"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int32"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceInfoDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceName": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "osName": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "deviceName",
          "deviceType"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_]{3,20}$"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{6,20}$"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfoDto"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ApiResponseTokenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/TokenResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64"
          },
          "refreshToken": {
            "type": "string"
          },
          "deviceToken": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "deviceInfo": {
            "$ref": "#/components/schemas/DeviceInfoDto"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
...
      "ApiResponseChunkResumeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ChunkResumeResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ChunkResumeResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "originalFileName": {
            "type": "string"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedChunks": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "pendingChunks": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "ApiResponseChunkProgressResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ChunkProgressResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ChunkProgressResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedChunks": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedBytes": {
            "type": "integer",
            "format": "int64"
          },
          "progress": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}