# src/main/java/com/evorsio/mybox/auth/dto/LoginRequest.java
package com.evorsio.mybox.auth.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {
    @NotBlank(message = "用户名不能为空")
    private String username;

    @NotBlank(message = "密码不能为空")
    private String password;

    // ===== 设备信息 =====
    private DeviceInfoDto deviceInfo;
}